Refactor HTTP helper exports
