+++ frontend/src/services/http/HttpHeaderCollection.ts
@@ -1,0 +1,3 @@
+export default interface HttpHeaderCollection {
+  [key: string]: string | undefined;
+}
