+++ frontend/src/services/http/HttpRequest.ts
@@ -1,0 +1,9 @@
+import HttpHeaderCollection from './HttpHeaderCollection';
+import URLQuery from './URLQuery';
+
+export default abstract class HttpRequest {
+  public constructor(
+    public readonly headers: HttpHeaderCollection = {},
+    public readonly query: URLQuery = {}
+  ) {}
+}
