+++ frontend/src/services/http/HttpServiceOptions.ts
@@ -1,10 +1,10 @@
 /**
  * Contains options for the {@link HttpService}
  */
-export interface HttpServiceOptions {
-    /**
-     * The base URI to use.
-     * If not specified, uses the base URI of the source the frontend was loaded from.
-     */
-    baseUri: string;
+export default interface HttpServiceOptions {
+  /**
+   * The base URI to use.
+   * If not specified, uses the base URI of the source the frontend was loaded from.
+   */
+  baseUri: string;
 }
