+++ frontend/src/services/http/URLQuery.ts
@@ -1,0 +1,5 @@
+export type URLQueryValueType = number | string | (number | string)[];
+
+export default interface URLQuery {
+  [key: string]: URLQueryValueType;
+}
